split out the tests
